+++ src/engine/Estimator.java
@@ -3,5 +3,5 @@ package engine;
 public abstract class Estimator { 
 	
 	abstract Frame calculateNextFrame (Frame frame);
-	abstract int calculateCompetingTags (Frame frame);
+	abstract double calculateCompetingTags (Frame frame);
 }
